v5.1

+++ src/index.ts
@@ -1,16 +1,16 @@
-import { MyDatabase } from "./db/database";
-import { Address, beginCell, TonClient, WalletContractV4 } from "@ton/ton";
-import { HIGHLOAD_CODE, highloadAddress, iotaEndpoint, rpcEndpoint, serviceChatID, tonApiEndpoint } from "./config";
+import {MyDatabase} from "./db/database";
+import {Address, beginCell, TonClient} from "@ton/ton";
+import {HIGHLOAD_CODE, highloadAddress, rpcEndpoint, serviceChatID, tonApiEndpoint} from "./config";
 import axios from "axios";
-import { Client } from "@iota/sdk";
-import { handleTransactions } from "./services/indexer/indexer";
-import { validateBalances } from "./services/validator/validator";
-import { configDotenv } from "dotenv";
-import { handleLiquidates } from "./services/liquidator";
-import { mnemonicToWalletKey } from "@ton/crypto";
-import { Bot } from "grammy";
+import {handleTransactions} from "./services/indexer/indexer";
+import {validateBalances} from "./services/validator/validator";
+import {configDotenv} from "dotenv";
+import {handleLiquidates} from "./services/liquidator";
+import {mnemonicToWalletKey} from "@ton/crypto";
+import {Bot} from "grammy";
 import * as https from "https";
-import { sleep } from "./helpers";
+import {sleep} from "./util/common";
+import {clearInterval} from "node:timers";
 
 async function main(bot: Bot) {
     configDotenv();
@@ -26,24 +26,12 @@ async function main(bot: Bot) {
             'Authorization': process.env.TONAPI_KEY
         }
     });
-    //mainnet
-    //const tonClient = new TonClient({
-    //    endpoint: rpcEndpoint,
-    //    apiKey: process.env.RPC_API_KEY
-    //});
-    //testnet
+
     const tonClient = new TonClient({
         endpoint: rpcEndpoint,
         apiKey: process.env.TONCENTER_API_KEY
     });
-    const iotaClient = new Client({
-        nodes: [iotaEndpoint],
-    });
     const keys = await mnemonicToWalletKey(process.env.WALLET_PRIVATE_KEY.split(' '));
-    const wallet = WalletContractV4.create({
-        workchain: 0,
-        publicKey: keys.publicKey
-    });
     const contract = tonClient.provider(Address.parse(highloadAddress), {
         code: HIGHLOAD_CODE,
         data: beginCell()
@@ -56,14 +44,18 @@ async function main(bot: Bot) {
 
     console.log(`Indexer is syncing...`);
     await handleTransactions(db, tonApi, tonClient, bot, Address.parse(highloadAddress), true);
-    console.log(`Indexer is synced. Waiting 5 sec before starting`);
+    console.log(`Indexer is synced. Waiting 1 sec before starting`);
 
-    await sleep(5000);
+    await sleep(1000);
+
     let handlingTransactions = false;
     const transactionID = setInterval(async () => {
-        if (handlingTransactions) return;
-        handlingTransactions = true;
+        if (handlingTransactions) {
+            console.log("[TransactionHandler]:", 'handling transactions in progress, wait more...');
+            return;
+        }
         console.log('Starting handleTransactions...')
+        handlingTransactions = true;
         handleTransactions(db, tonApi, tonClient, bot, Address.parse(highloadAddress))
             .catch(e => {
                 console.log(e);
@@ -83,8 +75,15 @@ async function main(bot: Bot) {
             });
     }, 5000);
 
+    let validating = false;
     const validatorID = setInterval(() => {
-        validateBalances(db, tonClient, iotaClient)
+        if (validating) {
+            console.log("[Validator]:", 'validation in progress, wait more...');
+            return;
+        }
+        validating = true;
+
+        validateBalances(db, tonClient, bot)
             .catch(e => {
                 console.log(e);
                 if (JSON.stringify(e).length == 2) {
@@ -93,8 +92,18 @@ async function main(bot: Bot) {
                 }
                 bot.api.sendMessage(serviceChatID, `[Validator]: ${JSON.stringify(e).slice(0, 300)}`);
             })
+            .finally(() => {
+                validating = false;
+            })
     }, 5000);
+
+    let liquidating = false;
     const liquidatorID = setInterval(() => {
+        if (liquidating) {
+            console.log("[Liquidator]:", 'liquidation in progress, wait more...');
+            return;
+        }
+        liquidating = true;
         handleLiquidates(db, tonClient, contract, Address.parse(highloadAddress), keys, bot)
             .catch(async (e) => {
                 console.log(e);
@@ -104,17 +113,49 @@ async function main(bot: Bot) {
                 }
                 await bot.api.sendMessage(serviceChatID, `[Liquidator]: ${JSON.stringify(e, null, 2).slice(0, 300)}`);
             })
-    }, 10000);
+            .finally(async () => {
+                liquidating = false;
+                console.log("Exiting from handleLiquidates...");
+            });
+    }, 5000);
 
-    setInterval(async () => {
-        const blacklistedUsers = await db.handleFailedTasks();
-        for (const user of blacklistedUsers) {
-            await bot.api.sendMessage(serviceChatID, `❌ User ${user} blacklisted`);
-            await sleep(100);
+    let blacklisting = false;
+    const blacklisterID = setInterval(async () => {
+        if (blacklisting) {
+            console.log("[Blacklister]:", "Blacklisting is in progress, wait more...");
+            return;
         }
+        blacklisting = true;
+        try {
+            const blacklistedUsers = await db.handleFailedTasks();
+            for (const user of blacklistedUsers) {
+                await bot.api.sendMessage(serviceChatID, `❌ User ${user} blacklisted`);
+                await sleep(100);
+            }
+            await db.deleteOldTasks();
+        } catch (e) {
+            console.log(e);
+        } finally {
+            blacklisting = false;
+            console.log("Exiting from blacklisting...");
+        }
+    }, 5000);
 
-        await db.deleteOldTasks();
-    }, 15000);
+    // handle interruption Ctrl+C === SIGINT
+    process.on('SIGINT', async () => {
+        clearInterval(transactionID);
+        clearInterval(validatorID);
+        clearInterval(liquidatorID);
+        clearInterval(blacklisterID);
+
+        const message = `Received SIGINT, stopping services...`;
+        console.log(message);
+        await bot.api.sendMessage(serviceChatID, message);
+
+        setTimeout(()=> {
+            throw ('Forced exit...');
+        }, 10_000);
+    });
 }
 
 (() => {
